feat: create source, operator and nodePorts file
