+++ packages/core/src/V2/nodes/link.ts
@@ -1,7 +1,7 @@
-import { PortProvider, IOperatorNodeConfig } from '../Node';
-import { CreateOutputPort, NodePorts, Operator } from './sleep';
+import { IOperatorNodeConfig, PortProvider } from '../Node';
 import { EMPTY, Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { Operator } from './operator';
+import { CreateOutputPort } from './nodePorts';
 
 export class LinkNodePorts implements PortProvider {
   constructor(private fromPort: Observable<unknown>, private toPortName: string) {
@@ -17,7 +17,7 @@ export class LinkNodePorts implements PortProvider {
 
 export const Link: IOperatorNodeConfig = {
   boot: (param: unknown) => {
-    const { from, to } = param as { from: string, to: string };
+    const { from, to } = param as {from: string, to: string};
     const createLinkOutput: CreateOutputPort = (input) => {
       return new LinkNodePorts(input.getPort(from), to);
     }
