+++ packages/core/src/V2/nodes/operator.ts
@@ -1,0 +1,15 @@
+import { OperatorNode, PortProvider } from '../Node';
+
+import { CreateOutputPort } from './nodePorts';
+
+export class Operator implements OperatorNode {
+
+  nodeType = 'operator' as const;
+
+  constructor(private createOutputPort: CreateOutputPort) {
+  }
+
+  getOutput(inputs: PortProvider): PortProvider {
+    return this.createOutputPort(inputs);
+  }
+}
